Refactor some of the date formatting and devotion models
