+++ iOS Client/Utilities/Formatters.swift
@@ -21,33 +21,6 @@ struct Formatters {
     }()
     
     static func getMonth(fromInt: Int) -> String {
-        switch fromInt {
-        case 1:
-            return "January"
-        case 2:
-            return "February"
-        case 3:
-            return "March"
-        case 4:
-            return "April"
-        case 5:
-            return "May"
-        case 6:
-            return "June"
-        case 7:
-            return "July"
-        case 8:
-            return "August"
-        case 9:
-            return "September"
-        case 10:
-            return "October"
-        case 11:
-            return "November"
-        case 12:
-            return "December"
-        default:
-            return ""
-        }
+        return Formatters.devotionDateFormatter.monthSymbols[fromInt-1]
     }
 }
